+++ Arduino_InvertedPendulum/KalmanFilter.h
@@ -1,12 +1,14 @@
+// Cf.: http://blog.tkjelectronics.dk/2012/09/a-practical-approach-to-kalman-filter-and-how-to-implement-it/
+
 class KalmanFilter
 {
-  private:
-    float angle;
-    float bias;
-    float P[2][2];
+private:
+  double angle;
+  double bias;
+  double P[2][2];
 
-  public:
-    KalmanFilter();
-    void setAngle(float newAngle);
-    float calcAngle(float newAngle, float newRate, float dt);
+public:
+  KalmanFilter();
+  void setAngle(double newAngle);
+  double calcAngle(double newAngle, double newRate, double dt);
 };
